% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/large.extractCoords.R
\name{cntrd}
\alias{cntrd}
\title{Find a centroid from long/lat matrix..}
\usage{
cntrd(mat, var.long, var.lat)
}
\arguments{
\item{mat}{a matrix-class object.}

\item{var.long}{Col.name}

\item{var.lat}{Col.name}
}
\description{
Find a centroid from long/lat matrix..
}
