% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/large.extractCoords.R
\name{large.extractCoords}
\alias{large.extractCoords}
\title{Extract coord matrix from SpatialPolygonDataframe and return the centroid of each block.}
\usage{
large.extractCoords(large.sp.df)
}
\arguments{
\item{large.sp.df}{a SpatialPolygonDataframe-class object.}
}
\description{
Extract coord matrix from SpatialPolygonDataframe and return the centroid of each block.
}
