% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loc_score.weight_mesh_val.R
\name{loc_score.weight_mesh_val}
\alias{loc_score.weight_mesh_val}
\title{Location score weighting by values on meshes.}
\usage{

  loc_score.weight_mesh_val(fn.SpatialPointDataFrame = "../Output/DataForMap.Wakayama_v01.RData",
  crs_default = "+init=epsg:4612", rank.restrict = 5,
  select_col.mesh_pop = c("PTC_2020", "PTD_2020", "PTE_2020"),
  fn.mesh.popEst = c("../Data/500m_mesh_suikei_2018_shape_24/500m_mesh_2018_24.shp",
  "../Data/500m_mesh_suikei_2018_shape_27/500m_mesh_2018_27.shp",
  "../Data/500m_mesh_suikei_2018_shape_29/500m_mesh_2018_29.shp",
  "../Data/500m_mesh_suikei_2018_shape_30/500m_mesh_2018_30.shp"),
  fn.Wakayama_Ph = "Wakayama_MasterAnaData.csv")
}
\arguments{
\item{fn.SpatialPointDataFrame}{RData file of SpatialPolygonDataFrame}

\item{crs_default}{projection.}

\item{rank.restrict}{Rank restriction. Default=5.}

\item{select_col.mesh_pop}{Column selector for mesh data (SpatialPolygonDataFrame@data$...)}

\item{fn.mesh.popEst}{Filename of mesh data.}

\item{fn.Wakayama_Ph}{Filename (.csv) of locations.}
}
\description{
Location score weighting by values on meshes.
}
