% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.loc_score.weight_mesh_val.R
\name{plot.wSDG}
\alias{plot.wSDG}
\title{Make histogram of diatance to pharmacy and each pharmacy's scores.}
\usage{
plot.wSDG(fn.RData.loc_score = "../Data/test_HmMs_2035.RData",
  rbPal = NULL, vec.prob_q = NULL, dir.Data = "../Data",
  weight.var_name = "score.merged_PTD_2035",
  Shape.pop_Est.var_name = "PTD_2035",
  fn.Shape.GovRegion = c("/190706/N03-190101_30_GML/N03-19_30_190101.shp",
  "/190706/N03-190101_24_GML/N03-19_24_190101.shp",
  "/190706/N03-190101_27_GML/N03-19_27_190101.shp",
  "/190706/N03-190101_29_GML/N03-19_29_190101.shp"),
  fn.ShapeP.SchoolRegion = "/190626/A32-16_30_GML/shape/A32P-16_30.shp",
  fn.Shape.SchoolRegion = "/190706/A32-13_30/shp/A32-13_30.shp",
  prefix.pdf_output = "location_score.only_pref_30.weight")
}
\arguments{
\item{fn.RData.loc_score}{= "../Data/test_HmMs_2035.RData",}

\item{rbPal}{= NULL,}

\item{vec.prob_q}{= NULL,}

\item{dir.Data}{= "../Data",}

\item{weight.var_name}{= "score.merged_PTD_2035",}

\item{Shape.pop_Est.var_name}{= "PTD_2035",}

\item{fn.Shape.GovRegion}{A character vector of (a) file name(s) (with ".shp") as the background.}

\item{fn.ShapeP.SchoolRegion}{= "/190626/A32-16_30_GML/shape/A32P-16_30.shp",}

\item{fn.Shape.SchoolRegion}{= "/190706/A32-13_30/shp/A32-13_30.shp",}

\item{prefix.pdf_output}{= "location_scor"}
}
\description{
Make histogram of diatance to pharmacy and each pharmacy's scores.
}
