% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/large.extractCoords.R
\name{extractCoords}
\alias{extractCoords}
\title{Extract coord matrix from SpatialPolygonDataframe@polygon (list-object).}
\usage{
extractCoords(polygons)
}
\arguments{
\item{polygons}{object.}
}
\description{
Extract coord matrix from SpatialPolygonDataframe@polygon (list-object).
}
